Default viewer backgrounds to transparent
